+++ CHANGES.md
@@ -1,9 +1,12 @@
-## Version 0.7.1 (in development)
+## Version 0.7.1 (from 2024-05-28)
 
-* Fixed link to _slice sources_ in documentation main page.
+* The function `zappend.api.zappend()` now returns the number of slices 
+  processed. (#93)
 
 * Moved all project configuration to `pyproject.toml` and removed
   `setup.cfg` and requirements files. (#88)
+
+* Fixed link to _slice sources_ in documentation main page.
 
 ## Version 0.7.0 (from 2024-03-19)